+++ packages/bundler/src/bundlers/esbuild/esbuild.config.ts
@@ -6,6 +6,8 @@ export default {
   platform: 'browser',
   mainFields: ['browser', 'module', 'main'],
   target: 'es2022',
+  keepNames: true,
+  allowOverwrite: true,
   loader: {
     '.js': 'js',
   },