+++ packages/bundler/src/polyfills/index.js
@@ -5,17 +5,23 @@ import fetchContext from './azion/fetch/index.js';
 import FirewallEventContext from './azion/firewall-event/index.js';
 import NetworkListContext from './azion/network-list/index.js';
 import { StorageContext } from './azion/storage/index.js';
+import cryptoContext from './crypto/index.js';
 import { fsContext } from './fs/index.js';
+import promisesContext from './promises/index.js';
+import { streamContext } from './stream/index.js';
 
 // TODO: transform polyfills to TypeScript
 
 export {
   AsyncHooksContext,
+  cryptoContext,
   EnvVarsContext,
   fetchContext,
   FetchEventContext,
   FirewallEventContext,
   fsContext,
   NetworkListContext,
+  promisesContext,
   StorageContext,
+  streamContext,
 };