+++ packages/bundler/src/polyfills/stream/context/index.js
@@ -1,0 +1,3 @@
+import streamContext from './stream.context.js';
+
+export default streamContext;