+++ packages/presets/src/presets/preact/index.ts
@@ -2,7 +2,6 @@ import type { AzionBuildPreset } from 'azion/config';
 import config from './config';
 import handler from './handler';
 import metadata from './metadata';
-// import prebuild from './prebuild';
-// import postbuild from './postbuild';
+import prebuild from './prebuild';
 
-export const preact: AzionBuildPreset = { config, metadata, handler };
+export const preact: AzionBuildPreset = { config, metadata, handler, prebuild };