+++ packages/unenv-preset/src/index.ts
@@ -5,7 +5,6 @@ const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 const getAbsolutePath = () => path.resolve(__dirname, '../../', 'unenv-preset', 'src');
 
-const nextNodePresetPath = `${getAbsolutePath()}/polyfills/node/frameworks/next`;
 const polyfillsPath = `${getAbsolutePath()}/polyfills`;
 
 export default {
@@ -13,34 +12,30 @@ export default {
     __dirname: `${polyfillsPath}/node/globals/path-dirname.js`,
     __filename: `${polyfillsPath}/node/globals/path-filename.js`,
     process: `${polyfillsPath}/node/globals/process.cjs`,
-    performance: `${polyfillsPath}/node/globals/performance.js`,
-    navigator: `${polyfillsPath}/node/globals/navigator.js`,
+    performance: `unenv/polyfill/performance`,
+    setInterval: `${polyfillsPath}/node/globals/set-interval.js`,
+    clearInterval: `${polyfillsPath}/node/globals/clear-interval.js`,
   },
   alias: {
     'azion/utils': 'azion/utils',
     'azion/utils/edge': 'azion/utils/edge',
     'azion/utils/node': 'azion/utils/node',
     '@fastly/http-compute-js': '@fastly/http-compute-js',
-    'next/dist/compiled/etag': `${nextNodePresetPath}/custom-server/12.3.x/util/etag.js`,
     accepts: 'accepts',
-    crypto: `${polyfillsPath}/node/crypto.js`,
-    events: 'events/events.js',
-    http: 'stream-http',
+    assert: 'assert-browserify',
+    https: `${polyfillsPath}/node/https.js`,
     module: `${polyfillsPath}/node/module.js`,
-    stream: 'stream-browserify/',
     string_decoder: 'string_decoder/lib/string_decoder.js',
-    url: 'url/url.js',
-    util: 'util/util.js',
     timers: 'timers-browserify/',
-    inherits: 'inherits/inherits_browser.js',
-    vm: 'vm-browserify/',
-    zlib: 'browserify-zlib',
+    util: `${polyfillsPath}/node/util.js`,
   },
-  external: ['node:async_hooks', 'node:fs/promises'],
+  external: ['node:async_hooks', 'node:fs/promises', 'node:stream', 'node:crypto'],
   polyfill: [
     'aziondev:async_hooks:/async-hooks/async-hooks.polyfills.js',
     'aziondev:fs:/fs/fs.polyfills.js',
     'aziondev:fs/promises:/fs/promises/promises.polyfills.js',
+    'aziondev:stream:/stream/stream.polyfills.js',
+    'aziondev:crypto:/crypto/crypto.polyfills.js',
     `azionprd:fs:/fs.js`,
   ],
 };