+++ packages/unenv-preset/src/polyfills/node/internal/util/inherits.js
@@ -1,0 +1,15 @@
+/* eslint-disable */
+export function inherits(ctor, superCtor) {
+  if (!superCtor) {
+    return;
+  }
+  ctor.super_ = superCtor;
+  ctor.prototype = Object.create(superCtor.prototype, {
+    constructor: {
+      value: ctor,
+      enumerable: false,
+      writable: true,
+      configurable: true,
+    },
+  });
+}